Commiting what I have for tonight. TXT files will NOT be local.

+++ TileRealms/Source/Enemy/Enemy.cs
@@ -14,7 +14,7 @@ namespace TileRealms
     {
         public Sprite sprite;
         public Vector2 location;
-        public double health;
+        public int health;
         EnemyController controller;
 
         public void Initialize(EnemyController e, Vector2 _location)
